--- Chrauma/Assets/Materials/Water.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6169b740e84193043b72f704104eaa1c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
