+++ Chrauma/Assets/Scenes/Lobby.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05828670bdf788c4f8215afe343a8cd4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
