+++ Chrauma/Assets/Scenes/Lobby/OcclusionCullingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2cff6b368fec08d4499170d28d58297f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 36300000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
