+++ Chrauma/Assets/Scenes/SceneDemo1/OcclusionCullingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0da81bc4e1e3f641af6bf67ca4579e9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 36300000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
