+++ Chrauma/Assets/Shaders/Water/Resources.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc77ad855be57214eb39fd06212f2ce4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
