+++ Chrauma/Assets/Shaders/Water/Resources/WaterReflection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfbe97460f8acd1488a82178305cdf59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
