+++ Chrauma/Assets/Shaders/Water/Resources/Water_mat_01.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 141d9d6ca43b6c843965491604c6deff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
