+++ Chrauma/Assets/Shaders/Water/Resources/Water_mat_03.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19ac4fcfcf42b4344a6c7d97a5cd61fc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
