--- Chrauma/Assets/Shaders/water/WaterBlock_50m.mesh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3da23cdf512599b42881988944f3918c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
