--- Chrauma/Assets/Shaders/water/WaterReflection.renderTexture.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4ae8ae7350b67b44785d874c413a2e03
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 8400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
