--- Chrauma/Assets/Shaders/water/WaterShader.shadergraph.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: dd97de70edf8a4141bee758a2efe7ff8
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
