--- Chrauma/Assets/Shaders/water/Water_mat_02.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8759ec410a6774b4ea13ae32dc2b8d2e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
