+++ Chrauma/Assets/player.save
@@ -1,11 +1,11 @@
 {
-    "health": 100,
+    "health": 175,
     "entropy": 100,
-    "scene": 1,
+    "scene": 2,
     "playerPos": {
-        "x": -121.53785705566406,
-        "y": 22.508865356445314,
-        "z": -16.273269653320314
+        "x": 82.32672119140625,
+        "y": -5.032066822052002,
+        "z": 4.230399131774902
     },
     "unlockable1": false,
     "unlockable2": false,
